refactor: export pull requests selector to dedicated component
