+++ src-vue/src/views/DeveloperView.vue
@@ -55,43 +55,9 @@
             <el-button type="primary" @click="clearFlightCorePersistentStore">
                 Delete FlightCore persistent store
             </el-button>
+
             <h3>Testing</h3>
-
-            <el-collapse>
-                <el-collapse-item title="Launcher PRs" name="1">
-                    <el-button type="primary" @click="getPullRequests('LAUNCHER')">
-                        Get launcher PRs
-                    </el-button>
-                    <p v-if="pull_requests_launcher.length === 0">No PRs loaded</p>
-                    <el-card v-else shadow="hover" v-for="pull_request in pull_requests_launcher"
-                        v-bind:key="pull_request.url">
-                        <el-button type="primary" @click="installLauncherPR(pull_request)">Install</el-button>
-                        <a target="_blank" :href="pull_request.html_url">
-                            {{ pull_request.number }}: {{ pull_request.title }}
-                        </a>
-                    </el-card>
-                </el-collapse-item>
-            </el-collapse>
-
-            <el-collapse>
-                <el-collapse-item title="Mods PRs" name="1">
-                    <el-alert title="Warning" type="warning" :closable="false" show-icon>
-                        Mod PRs are installed into a separate profile. Make sure to launch via
-                        'r2ns-launch-mod-pr-version.bat' or via '-profile=R2Northstar-PR-test-managed-folder' to actually
-                        run the PR version!
-                    </el-alert>
-                    <el-button type="primary" @click="getPullRequests('MODS')">
-                        Get Mods PRs
-                    </el-button>
-                    <p v-if="pull_requests_mods.length === 0">No PRs loaded</p>
-                    <el-card v-else shadow="hover" v-for="pull_request in pull_requests_mods" v-bind:key="pull_request.url">
-                        <el-button type="primary" @click="installModsPR(pull_request)">Install</el-button>
-                        <a target="_blank" :href="pull_request.html_url">
-                            {{ pull_request.number }}: {{ pull_request.title }}
-                        </a>
-                    </el-card>
-                </el-collapse-item>
-            </el-collapse>
+            <pull-requests-selector />
         </el-scrollbar>
     </div>
 </template>
@@ -102,20 +68,14 @@ import { invoke } from "@tauri-apps/api";
 import { ElNotification } from "element-plus";
 import { GameInstall } from "../utils/GameInstall";
 import { Store } from 'tauri-plugin-store-api';
-import { PullsApiResponseElement } from "../../../src-tauri/bindings/PullsApiResponseElement";
-import { PullRequestType } from "../../../src-tauri/bindings/PullRequestType";
 import { ReleaseCanal } from "../utils/ReleaseCanal";
+import PullRequestsSelector from "../components/PullRequestsSelector.vue";
 const persistentStore = new Store('flight-core-settings.json');
 
 export default defineComponent({
     name: "DeveloperView",
-    computed: {
-        pull_requests_launcher(): PullsApiResponseElement[] {
-            return this.$store.state.pull_requests_launcher;
-        },
-        pull_requests_mods(): PullsApiResponseElement[] {
-            return this.$store.state.pull_requests_mods;
-        },
+    components: {
+        PullRequestsSelector
     },
     data() {
         return {
@@ -259,82 +219,6 @@ export default defineComponent({
             await persistentStore.clear();
             // ...and save
             await persistentStore.save();
-        },
-        async getPullRequests(pull_request_type: PullRequestType) {
-            await invoke<PullsApiResponseElement[]>("get_pull_requests_wrapper", { installType: pull_request_type })
-                .then((message) => {
-                    console.log(message);
-                    // Show user notification if mod install completed.
-                    ElNotification({
-                        title: `Done`,
-                        message: `Loaded pull requests`,
-                        type: 'success',
-                        position: 'bottom-right'
-                    });
-
-                    switch (pull_request_type) {
-                        case "MODS":
-                            this.$store.state.pull_requests_mods = message;
-                            break;
-
-                        case "LAUNCHER":
-                            this.$store.state.pull_requests_launcher = message;
-                            break;
-
-                        default:
-                            console.error("We should never end up here");
-                    }
-                })
-                .catch((error) => {
-                    ElNotification({
-                        title: 'Error',
-                        message: error,
-                        type: 'error',
-                        position: 'bottom-right'
-                    });
-                });
-        },
-        async installLauncherPR(pull_request: PullsApiResponseElement) {
-            await invoke("apply_launcher_pr", { pullRequest: pull_request, gameInstallPath: this.$store.state.game_path })
-                .then((message) => {
-                    console.log(message);
-                    // Show user notification if mod install completed.
-                    ElNotification({
-                        title: `Done`,
-                        message: `Installed ${pull_request.number}: "${pull_request.title}"`,
-                        type: 'success',
-                        position: 'bottom-right'
-                    });
-                })
-                .catch((error) => {
-                    ElNotification({
-                        title: 'Error',
-                        message: error,
-                        type: 'error',
-                        position: 'bottom-right'
-                    });
-                });
-        },
-        async installModsPR(pull_request: PullsApiResponseElement) {
-            console.log(pull_request);
-            await invoke("apply_mods_pr", { pullRequest: pull_request, gameInstallPath: this.$store.state.game_path })
-                .then((message) => {
-                    // Show user notification if mod install completed.
-                    ElNotification({
-                        title: `Done`,
-                        message: `Installed ${pull_request.number}: "${pull_request.title}"\nMake sure to launch via batch file or by specifying correct profile!`,
-                        type: 'success',
-                        position: 'bottom-right'
-                    });
-                })
-                .catch((error) => {
-                    ElNotification({
-                        title: 'Error',
-                        message: error,
-                        type: 'error',
-                        position: 'bottom-right'
-                    });
-                });
         },
         async forceInstallNorthstar() {
             let game_install = {
